refactor: move the constants and functions to different files, change the utils folder to the libs folder and rename the constants
