+++ currency-converter/src/components/AddListCurrency/AddListCurrency.tsx
@@ -1,12 +1,13 @@
 import { List, ListItem, ListItemText } from "@mui/material";
 import { useContext } from "react";
 import { IListCurrencies } from "../../types/types";
-import { Context, saveCurrencyToList } from "../../utils/utils";
+import { saveCurrencyToList } from "../../libs/currency";
+import { CONTEXT } from "../../libs/constants";
 import styles from "./AddListCurrency.module.scss";
 
 function AddListCurrency(props: IListCurrencies) {
   const { listCurrencies, setListCurrencies } = props;
-  const { currenciesAllNames } = useContext(Context);
+  const { currenciesAllNames } = useContext(CONTEXT);
 
   function addToList(nameCurrency: string) {
     saveCurrencyToList(listCurrencies, setListCurrencies, nameCurrency);
