+++ currency-converter/src/hooks/useCurrencies.ts
@@ -1,10 +1,10 @@
 import { useMemo } from "react";
 import { ICurrency } from "../types/types";
-import { favouriteProperty } from "../utils/utils";
+import { FAVOURITE_PROPERTY } from "../libs/constants";
 
 export const useSortedCurrencies = (currencies: ICurrency[], type: string) => {
   const sortedCurrencies = useMemo(() => {
-    if (type === favouriteProperty) {
+    if (type === FAVOURITE_PROPERTY) {
       return [...currencies].sort((x, y) => {
         return Number(y.favourite) - Number(x.favourite);
       });
